requirements.txt
# Core LangChain
langchain==0.1.20 
langchain-community==0.0.38 
langchain-core==0.1.52

# Vector Store & Retrieval
faiss-cpu
rank-bm25

# LLM Integration
requests==2.31.0

# Data Processing
pydantic==2.5.3

# Utilities
python-dotenv==1.0.0

# UI (Optional)
streamlit==1.29.0

# Environment
python-dotenv==1.0.1